Redo terminal colours for vrunchbang scheme
Terminal colours in the scheme were totally off (meant for the dark
version?). Replaced all of them with the closest matches for the
corresponding GUI 24-bit colours. The matching was done using this [1]
script.
[1] https://gist.github.com/MicahElliott/719710

The resulting approximation of the colour scheme for a 256-colour
terminal looks very descent. This allows to make the setting of
truecolour optional.