Set xterm mouse handling explicitly
If Vim supports protocol SGR, set the terminal type to it. This or at
least xterm2 is needed to report mouse dragging, which then enables
resizing of split panels with mouse.

This is a work-around needed when Vim is used from within GNU screen.
Apparently, screen does not propagate the xterm version properly.